requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-ollama==0.2.2
langgraph==0.2.60
pydantic==2.10.4
python-dotenv==1.0.1
requests==2.32.3

